# cpu: aarch64: remove unnecessary workaround for f16 eltwise_tanh